refactor: replaced component imports and deleted old component folder
