+++ src/Components/GuildCard/GuildCardContent.tsx
