+++ src/Components/GuildCard/types.tsx
@@ -3,4 +3,3 @@ import { CardProps } from 'old-components/Guilds/common/Card/index';
 export interface GuildCardProps extends CardProps {
   guildAddress: string;
 }
-
